refactor: Separate components
